Minor bug fix
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	modified:   .gitignore
	modified:   server.py
	modified:   serverinfoparser.py

fixed unexpected call of onPlayerJoin with player "*" when a player
inputs "/me chino joined the game"
fixed an unimportant bug: for example the input
"[Server thread/INFO]: !!tps" will work like "!!tps".
And some minor change to the code.